fixed with stimulus on connect
