+++ app/javascript/controllers/flatpickr_service_form_controller.js
@@ -1,0 +1,14 @@
+import { Controller } from "@hotwired/stimulus"
+import flatpickr from "flatpickr"
+
+export default class extends Controller {
+  static targets = ["input"]
+
+  connect() {
+    flatpickr(this.inputTarget, {
+        minDate: "today",
+        mode: "multiple",
+        dateFormat: "Y-m-d",
+    });
+  }
+}
