+++ app/javascript/packs/booking_new.js
@@ -1,12 +1,12 @@
-import flatpickr from "flatpickr"
+// import flatpickr from "flatpickr"
 
 
-const availableDates = JSON.parse(document.querySelector('#booking-form').dataset.available)
-const dateInput = document.getElementById("booking_date")
-if (dateInput !== null) {
-  flatpickr(dateInput, {
-      minDate: "today",
-      enable: availableDates,
-      dateFormat: "Y-m-d",
-  });
-}
+// const availableDates = JSON.parse(document.querySelector('#booking-form').dataset.available)
+// const dateInput = document.getElementById("booking_date")
+// if (dateInput !== null) {
+//   flatpickr(dateInput, {
+//       minDate: "today",
+//       enable: availableDates,
+//       dateFormat: "Y-m-d",
+//   });
+// }
