+++ app/views/bookings/new.html.erb
@@ -1,7 +1,7 @@
-<%= javascript_pack_tag 'booking_new' %>
+<%# <%= javascript_pack_tag 'booking_new' %>
 <h1> Your booking </h1>
 <div class="container col-6">
-  <div id='booking-form' data-available="<%= @service.availability.to_json %>">
+  <div id='booking-form' data-controller="flatpickr-booking-form" data-available="<%= @service.availability.to_json %>">
     <%= simple_form_for [@service, @booking] do |f| %>
       <%= f.input :vehicle_brand %>
       <%= f.input :vehicle_model %>
@@ -9,7 +9,7 @@
       <%= f.input :vehicle_year %>
       <%= f.input :vehicle_contact %>
       <%= f.input :additional_details, as: :text %>
-      <%= f.input :date, as: :date, html5: true %>
+      <%= f.input :date, as: :date, html5: true, input_html: { data: { flatpickr_booking_form_target: "input"}} %>
       <%= f.submit %>
     <% end %>
   </div>
